requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
joblib==1.3.2
xgboost==1.7.6
lightgbm==4.0.0
tensorflow==2.13.0
keras==2.13.1
pyserial==3.5
pymongo==4.4.1
flask==2.3.2
flask-cors==4.0.0
requests==2.31.0
schedule==1.2.0
python-dotenv==1.0.0 